+++ src/components/Playground/ImportButton.scss
@@ -1,0 +1,13 @@
+.btn {
+  &.import {
+    background-color: var(--panel-background-active-color);
+    color: var(--foreground-color);
+    border: 0;
+    //border-radius: 4px;
+    //cursor: pointer;
+
+    &:hover {
+      background-color: var(--background-color);
+    }
+  }
+}
